+++ core/src/main/java/com/github/kongchen/swagger/docgen/DocTemplateConstants.java
