+++ core/src/main/java/com/github/kongchen/swagger/docgen/mustache/MustacheAuthorization.java
