+++ core/src/main/java/com/github/kongchen/swagger/docgen/mustache/MustacheResponseClass.java
