+++ core/src/main/java/com/github/kongchen/swagger/docgen/util/Utils.java
