+++ mavenplugin/src/main/java/com/github/kongchen/swagger/docgen/mavenplugin/IncludeProjectDependenciesComponentConfigurator.java
