--- src/test/java/issue17/Child.java
@@ -1,18 +1,0 @@
-package issue17;
-
-/**
- * Created with IntelliJ IDEA.
- * User: kongchen
- * Date: 7/31/13
- */
-public class Child extends Parent{
-    private String name;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
