--- src/test/java/sample/model/G1.java
@@ -1,25 +1,0 @@
-package sample.model;
-
-/**
- * Created by chekong on 13-6-9.
- */
-public class G1 {
-    private String g1;
-    private int a1;
-
-    public String getG1() {
-        return g1;
-    }
-
-    public void setG1(String g1) {
-        this.g1 = g1;
-    }
-
-    public int getA1() {
-        return a1;
-    }
-
-    public void setA1(int a1) {
-        this.a1 = a1;
-    }
-}
