--- src/test/java/sample/model/G2.java
@@ -1,17 +1,0 @@
-package sample.model;
-
-/**
- * Created by chekong on 13-6-9.
- */
-public class G2 {
-    private String a2;
-    private int g2;
-
-    public String getA2() {
-        return a2;
-    }
-
-    public void setA2(String a2) {
-        this.a2 = a2;
-    }
-}
